A ggplot2 labeller for FLQuants

Plots of FLQuants objects use by default facet_grid to separate the different elements. This function generates facet labels that have both the element name and the units of measurement of each, as stored in the 'units' slot of each FLQuant.

label_flqs(x, drop = c("NA", "NC", "m", "f", "z", "prop"))

Arguments

x

An object of class FLQuants

drop

Character string to be dropped from the label shen found in the *units* slot. Defaults to c("NA", "NC", "m", "z", "prop")

Value

A ggplot2 labeller function

Details

Certain *units* are dropped from the label, as being uninformative: "NA", "NC", "m", "z", and "prop". This can be selected with the *drop* argument.

See also

labeller