A high-contrast palette to use in ggplotFL

Plot methods defined in the ggplotFL package make use by default of a palette with a high contrast, useful to separate time series or categories. The palette consist of seven colours: red, blue, green, violet, orange, yellow and brown. This palette can be inspected at the [colorbrewe2.org site](http://colorbrewer2.org/?type=qualitative&scheme=Set1&n=7#type=qualitative&scheme=Set1&n=7).

flpalette

flpalette_colours(n = length(flpalette))

flpalette_grads(palette = flpalette, reverse = FALSE, ...)

scale_colour_flr(palette = flpalette, discrete = TRUE, reverse = FALSE, ...)

Arguments

n
Number of colours or individual colours to return, or number of colours to interpolate.

palette
Palette subset to create a gradual scale from, defaults to *flpalette*.

reverse
Should the palette be reversed, FALSE.

...
Elements to subset from palette, by name or position.

discrete
Is the palette to be applied to a discrete variable, TRUE.

Format

An object of class character of length 7.

Value

A named vector of colors and HEX codes, or a function to obtain a gradient of colors fo a given length.

Details

The palette is accessible as a named vector, *flpalette*. Two functions are also available to manipualte the palette. One to extract a subset of the palette, *flpalette_colours*, and another to create a gradation of colours between two or more of the palette colours, *flpalette_grads*.

See also

Author

The FLR Team

Examples

# CHECK flpalette
 flpalette

#>       red      blue     green    violet    orange    yellow     brown 
#> "#e41a1c" "#377eb8" "#4daf4a" "#984ea3" "#ff7f00" "#ffff33" "#a65628" 

flpalette_colours()

#> [1] "#e41a1c" "#377eb8" "#4daf4a" "#984ea3" "#ff7f00" "#ffff33" "#a65628"

flpalette_colours(5)

#> [1] "#e41a1c" "#377eb8" "#4daf4a" "#984ea3" "#ff7f00"

flpalette_colours(2:3)

#> [1] "#377eb8" "#4daf4a"

flpalette_grads(flpalette_colours(3))(20)

#>  [1] "#E41A1C" "#D1242C" "#BF2F3C" "#AD394D" "#9B445D" "#884E6E" "#76597E"
#>  [8] "#64638E" "#526E9F" "#4078AF" "#3880B2" "#3A85A6" "#3C8A9B" "#3F908F"
#> [15] "#419583" "#439A78" "#469F6C" "#48A461" "#4AA955" "#4DAF4A"